refactor(stylex_shared): move nonstandard polyfill tests to legacy path
