+++ crates/stylex-shared/tests/__swc_snapshots__/tests/legacy/transform_legacy_polyfills_test/css_property_polyfills.rs/non_standard_padding_horizontal_aka_padding_inline.js
